Refactor mutators to use a unified createMutators function and update imports to align with the new package structure. Remove deprecated example and mutators files.
