--- packages/zero/server.ts
@@ -1,1 +1,0 @@
-export * from '@openfaith/zero-effect/zeroLayer'
